requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.8.2
python-dotenv==1.0.1
pymysql==1.1.0
httpx==0.27.0
openai==1.51.0
psycopg[binary]==3.1.18
PyJWT==2.8.0
bcrypt==4.1.2
google-api-python-client==2.108.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.1.0
pytz==2023.3